task #3: create functional for profile page
