+++ src/components/app-header/app-header.jsx
@@ -1,13 +1,21 @@
 import React from 'react';
+import { useNavigate } from "react-router-dom";
+
 import appHeaderStyles from './app-header.module.css';
+
 import { Logo, BurgerIcon, ListIcon, ProfileIcon } from '@ya.praktikum/react-developer-burger-ui-components'
 
 function AppHeader() {
+  const navigate = useNavigate();
+
   const renderComponents = [
     {
       text: 'Конструктор',
       icon: BurgerIcon,
       id: 'burger',
+      handler: () => {
+        navigate('');
+      },
     },
     {
       text: 'Лента заказов',
@@ -18,8 +26,17 @@ function AppHeader() {
       text: 'Личный кабинет',
       icon: ProfileIcon,
       id: 'account',
+      handler: () => {
+        navigate('/profile');
+      },
     },
   ];
+
+  const headerBtnsHandler = (btn) => {
+    if (btn.handler) {
+      btn.handler();
+    }
+  };
 
   return (
     <header
@@ -29,6 +46,7 @@ function AppHeader() {
         {
           renderComponents.map((btn) => (
             <button
+            onClick={() => {headerBtnsHandler(btn)}}
             key={btn.id}
             className={`mr-2 pr-5 pl-5 ${appHeaderStyles['app-header__btn']}`}>
               {
